requirements-api.txt
langchain
openai
elasticsearch
pydantic
python-dotenv
numpy
pandas
fastapi
uvicorn
python-multipart
pdfplumber
pypdf
tiktoken
aiohttp
PyYAML
pytesseract
pdf2image
pypandoc
pybase64
python-csv
psycopg2-binary
PyMuPDF
pymongo
networkx
openpyxl
odfpy
olefile
python-docx
httpx
sentence-transformers
nltk
torch
bson
filetype
markdown
mistralai
neo4j
pillow-heif
watchdog
PyJWT
bcrypt
email-validator
